add log for napi

+++ service/event_handler/src/key_event_value_transformation.cpp
@@ -513,7 +513,8 @@ int32_t KeyItemsTransKeyIntention(const std::vector<KeyEvent::KeyItem> &items)
 
     int64_t keyCodes = 0;
     for (const auto &item : items) {
-        keyCodes = static_cast<int64_t>((static_cast<uint64_t>(keyCodes) << BIT_SET_INDEX) + item.GetKeyCode());
+        keyCodes = static_cast<int64_t>(
+            (static_cast<uint64_t>(keyCodes) << BIT_SET_INDEX) + (static_cast<uint64_t>(item.GetKeyCode())));
     }
     auto iter = MAP_KEY_INTENTION.find(keyCodes);
     if (iter == MAP_KEY_INTENTION.end()) {
